add DummyFileResolver to ruff_db
initially I was going to add this directly in the formatter test fixture file,
but we would have to import NotebookIndex from somewhere, so it seemed easiest
just to drop this directly in the render module. we could possibly reuse this
elsewhere, like in the parser fixtures, which currently have their own syntax
error rendering code.