[ty] Mark completion responses as "incomplete"
This means that when new characters are typed, the LSP client should
send new completion requests to the LSP server. Previously, only one
request would be emitted and then the client would filter them.

The primary reason to do this is so the server can take what has been
typed into account when returning completions. ty is already doing this
with auto-import, because it isn't really feasible to do otherwise:
always returning *every* symbol throughout an environment isn't a good
way to do business. But, we likely also want to take what is typed into
account for all completion requests so that we can do our own ranking.

Moreover, if we take what is typed into account when doing a completion
request, then consider a request that was initiated with `fubar` already
typed. If the user hits backspace and gets `fub`, then since the LSP
client isn't sending another request, the user won't see all matching
completions.

This is why we need to mark the response as "incomplete." This will
instruct the LSP client to re-send completion requests.

The trade-off is that the client will make more completion requests. And
if completions are slow, this will magnify that effect for users.