[ruff] Expose more information about tokens
For completions, I want to be able to analyze the
token stream to determine, somewhat reliably, if
the cursor is within an unclosed string. I didn't
see a way to do that easily without exposing some
of the token flags.

One idiosyncracy here is that `TokenFlags` seems
like it's only supposed to be used when you *know*
a token is a string. But I want to be able to inspect
it for a "unknown" token in the case of an unclosed
string. In that case, an unknown token for an unclosed
single quoted string has no discerning characteristics.
Hence, I added a new `SINGLE_QUOTES` flag here.

I think this is non-ideal for at least two reasons.

Firstly, it increases the size of `TokenFlags` to a
`u16`. And this is done even though this is not needed
by the lexer.

Secondly, currently, it is *assumed* that if there
isn't `DOUBLE_QUOTES` then it must imply that the string
is `SINGLE_QUOTES`. And this works because `TokenFlags`
is only inspected when the token is known to be a string
(and never when it is unknown).

Suggestions on how to improve this are welcome, but this
seemed like a plausible route to me.