[ty] Don't show completions in strings
... except when we're in the interpolated segments of an f-string or a
t-string.

Previously, we would show completions for interpolated segments of an
f-string or a t-string. Which is good. But we did this by virtue of
never trying to suppress completions when in the non-interpolated part
of an f-string/t-string, or any other kind of string. Whcih is bad.

Here, we do our best to detect whether the cursor is within a string or
not. This seems somewhat straight-forward in the case of a string that
is properly opened/closed. But for a string that isn't closed
yet---which is likely common while typing---this is less
straight-forward. For the most part, I looked at some examples and tried
to reverse engineer the kinds of patterns we see in the token stream
when there is an unclosed string. It's likely not fully correct, but we
can improve it as we go.