Add version, file, and line metadata to `Violation` structs
Summary
--

This PR adds the version a rule was added, and the file and line where it's
defined to `ViolationMetadata`. The file and line just use the standard `file!`
and `line!` macros, while the more interesting version field uses a new
`violation_metadata(version = ...)` attribute parsed by our `ViolationMetadata`
derive macro. I'm not sure that this is the best way to go, but it seemed like a
reasonable start.

I've also included the scripts I used to generate and add these in the PR for
now. I don't really think we need to keep them around permanently, but it seemed
like they could be helpful if anything looks wrong. I had Claude generate the
first iteration of `generate_rule_metadata.py`, which hopefully explains all of
the emojis, but I had to modify it pretty heavily today to get it working
correctly. I think `run_uvx_rule_check` and `get_ruff_version` are the most
interesting parts. The first because it shows ways to check rule availability
across a wide range of Ruff versions, and the latter because it filters out some
versions that uv couldn't install.

As a curiosity and a bit of a sanity check, I also plotted the rule numbers over
time:

TODO

I think this looks pretty reasonable and avoids some of the artifacts the
earlier versions of the script ran into, such as the `rule` sub-command not
being available or `--explain` requiring a file argument.

<details><summary>Script and summary data</summary>

```shell
gawk --csv '
NR > 1 {
    split($2, a, ".")
    major = a[1]; minor = a[2]; micro = a[3]
    # sum the number of rules added per minor version
    versions[minor] += 1
}
END {
    tot = 0
    for (i = 0; i <= 14; i++) {
        tot += versions[i]
        print i, tot
    }
}
' ruff_rules_metadata.csv > summary.dat
```

```
0 696
1 768
2 778
3 803
4 822
5 848
6 855
7 865
8 893
9 915
10 916
11 924
12 929
13 932
14 933
```

</details>

Test Plan
--

I temporarily modified the `fix_title` of one of the rules to print the version,
file, and line number, just to make sure the macro was working. The real test
will be using this as part of our documentation generation, which I can either
add on here or in an immediate follow-up PR.